adapted wxSizer::Remove type to new header
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@18772 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ src/generic/proplist.cpp
@@ -542,7 +542,7 @@ void wxPropertyListView::ShowListBoxControl(bool show)
         if (show)
             m_middleSizer->Prepend( m_valueList, 0, wxTOP|wxLEFT|wxRIGHT | wxEXPAND, 3 );
         else
-            m_middleSizer->Remove( 0u );
+            m_middleSizer->Remove( (size_t) 0 );
 
         m_propertyWindow->Layout();
     }
